CO2 concnetration

CO2 concentration used to make the plot in the poster

Usage

conc_poster

Format

A tibble with 530 rows and 11 variables

datetime

Datetime at which CO2 concentration was recorded.

f_conc

CO2 concentration in ppm.

f_start

Datetime at which the measurement was started.

f_fluxID

Unique ID for each flux.

f_cut

Indicating if the measurement should be kept (keep) or discarded (cut).

f_b

b parameter of the C(t) function, calculated by optim() with b_est as starting point.

f_fit

C(t), modeled CO2 concentration as a function of time.

fit_slope

Output of linear model of CO2 concentration passing by C(tz) and a slope of slope_tz.

f_cor_coef

coefficient of correlation between gas concentration and time

f_RMSE

RMSE of the exponential fit and the measured data

f_quality_flag

quality flag advising if the slope has to be replaced by 0 or NA

Examples

conc_poster
#> # A tibble: 530 × 11
#>    f_datetime          f_conc f_fluxID f_quality_flag f_start             f_RMSE
#>    <dttm>               <dbl> <fct>    <chr>          <dttm>               <dbl>
#>  1 2020-07-17 16:49:55   460. 37       discard        2020-07-17 16:49:55   26.1
#>  2 2020-07-17 16:49:56   519. 37       discard        2020-07-17 16:49:55   26.1
#>  3 2020-07-17 16:49:57   509. 37       discard        2020-07-17 16:49:55   26.1
#>  4 2020-07-17 16:49:58   471. 37       discard        2020-07-17 16:49:55   26.1
#>  5 2020-07-17 16:49:59   446. 37       discard        2020-07-17 16:49:55   26.1
#>  6 2020-07-17 16:50:00   431. 37       discard        2020-07-17 16:49:55   26.1
#>  7 2020-07-17 16:50:01   424. 37       discard        2020-07-17 16:49:55   26.1
#>  8 2020-07-17 16:50:02   420. 37       discard        2020-07-17 16:49:55   26.1
#>  9 2020-07-17 16:50:03   419. 37       discard        2020-07-17 16:49:55   26.1
#> 10 2020-07-17 16:50:04   417. 37       discard        2020-07-17 16:49:55   26.1
#> # ℹ 520 more rows
#> # ℹ 5 more variables: f_cor_coef <dbl>, f_b <dbl>, f_cut <fct>, f_fit <dbl>,
#> #   f_fit_slope <dbl>