calculates fluxes

calculates a flux based on the rate of change of gas concentration over time

Usage

flux_calc(
  slope_df,
  slope_col,
  cut_col = c(),
  keep_filter = c(),
  chamber_volume = 24.5,
  tube_volume = 0.075,
  atm_pressure = 1,
  plot_area = 0.0625,
  R_const = 0.082057,
  cols_keep = c(),
  cols_ave = c(),
  fluxID_col = "f_fluxID",
  temp_air_col = "temp_air",
  temp_air_unit = "celsius"
)

Arguments

slope_df

dataframe of flux slopes

slope_col

column containing the slope to calculate the flux

cut_col

column containing cutting information

keep_filter

name in cut_col of data to keep

chamber_volume

volume of the flux chamber in L, default for Three-D project chamber (25x24.5x40cm), can also be a column in case it is a variable

tube_volume

volume of the tubing in L, default for summer 2020 setup, can also be a column in case it is a variable

atm_pressure

atmoshperic pressure, assumed 1 atm, can be a constant (numerical) or a variable (column name)

plot_area

area of the plot in m^2, default for Three-D

R_const

gas constant (0.082057 LatmK^(-1)*mol^(-1))

cols_keep

columns to keep from the input to the output. Those columns need to have unique values for each flux.

cols_ave

columns with values that should be averaged for each flux in the ouput. Note that NA are removed in mean calculation

fluxID_col

column containing the fluxID

temp_air_col

column containing the air temperature used to caculate fluxes. Will be averaged with NA removed.

temp_air_unit

units in which air temperature was measured. Has to be either celsius, fahrenheit or kelvin

Value

a df containing fluxID, fluxes (in mmol*m^(-2)*h^(-1)), temperature average for each flux, slope used for each flux calculation, and any columns specified in cols_keep and cols_ave.

Examples

data(slopes0)
flux_calc(slopes0, slope_col = "f_slope_tz")
#> # A tibble: 6 × 8
#>   f_fluxID f_slope_calc chamber_volume tube_volume atm_pressure temp_air_ave
#>      <dbl>        <dbl>          <dbl>       <dbl>        <dbl>        <dbl>
#> 1        1        1.56            24.5       0.075            1         7.31
#> 2        2        0.853           24.5       0.075            1         7.38
#> 3        3        0.303           24.5       0.075            1         7.46
#> 4        4        1.13            24.5       0.075            1         7.77
#> 5        5        1.46            24.5       0.075            1         7.71
#> 6        6        0.426           24.5       0.075            1         7.75
#> # ℹ 2 more variables: volume_setup <dbl>, flux <dbl>