Fitting a model to the gas concentration curve and estimating the slope

Fits an exponential expression to the concentration evolution over time, as described in Zhao 2018

Usage

flux_fitting_exp(
  conc_df,
  t_window = 20,
  Cz_window = 15,
  b_window = 10,
  a_window = 10,
  roll_width = 15,
  start_cut = 0,
  end_cut = 0,
  start_col = "start",
  end_col = "end",
  datetime_col = "datetime",
  conc_col = "conc",
  fluxID_col = "fluxID"
)

Arguments

conc_df

dataframe of gas concentration over time

t_window

enlarge focus window before and after tmin and tmax

Cz_window

window used to calculate Cz, at the beginning of cut window

b_window

window to estimate b. It is an interval after tz where it is assumed that C fits the data perfectly

a_window

window at the end of the flux to estimate a

roll_width

width of the rolling mean for CO2 when looking for tz, ideally same as Cz_window

start_cut

to cut at the start

end_cut

to cut at the end, if you notice on the plots that the match was not precise enough

start_col

column with datetime when the measurement started

end_col

column with datetime when the measurement ended

datetime_col

column with datetime of each concentration measurement

conc_col

column with gas concentration data

fluxID_col

column with ID of each flux

Value

a dataframe with the slope at t zero, modelled concentration over time and exponential expression parameters

Examples

data(co2_conc)
flux_fitting_exp(co2_conc)
#> Joining with `by = join_by(fluxID)`
#> Joining with `by = join_by(fluxID)`
#> Joining with `by = join_by(fluxID)`
#> Joining with `by = join_by(fluxID)`
#> Joining with `by = join_by(fluxID)`
#> Joining with `by = join_by(fluxID)`
#> Joining with `by = join_by(fluxID)`
#> Joining with `by = join_by(fluxID)`
#> Joining with `by = join_by(fluxID)`
#> Joining with `by = join_by(datetime, temp_air, temp_soil, conc, PAR, turfID,
#> type, start, end, fluxID, n_conc, ratio, flag, time, cut)`
#> Warning: 
#>  fluxID 5 : slope was estimated on 205 points out of 210 seconds because data are missing
#>  fluxID 6 : slope was estimated on 206 points out of 210 seconds because data are missing
#> # A tibble: 1,251 × 29
#>    datetime            temp_air temp_soil  conc   PAR turfID       type 
#>    <dttm>                 <dbl>     <dbl> <dbl> <dbl> <fct>        <fct>
#>  1 2022-07-28 23:43:35    NA         NA    447. NA    156 AN2C 156 ER   
#>  2 2022-07-28 23:43:36     7.22      10.9  447.  1.68 156 AN2C 156 ER   
#>  3 2022-07-28 23:43:37    NA         NA    448. NA    156 AN2C 156 ER   
#>  4 2022-07-28 23:43:38    NA         NA    449. NA    156 AN2C 156 ER   
#>  5 2022-07-28 23:43:39    NA         NA    449. NA    156 AN2C 156 ER   
#>  6 2022-07-28 23:43:40    NA         NA    450. NA    156 AN2C 156 ER   
#>  7 2022-07-28 23:43:41    NA         NA    451. NA    156 AN2C 156 ER   
#>  8 2022-07-28 23:43:42    NA         NA    451. NA    156 AN2C 156 ER   
#>  9 2022-07-28 23:43:43    NA         NA    453. NA    156 AN2C 156 ER   
#> 10 2022-07-28 23:43:44    NA         NA    453. NA    156 AN2C 156 ER   
#> # ℹ 1,241 more rows
#> # ℹ 22 more variables: start <dttm>, end <dttm>, fluxID <fct>, n_conc <int>,
#> #   ratio <dbl>, flag <chr>, time <dbl>, cut <fct>, Cm_est <dbl>, a_est <dbl>,
#> #   b_est <dbl>, tz_est <dbl>, Cz <dbl>, time_diff <dbl>, Cm <dbl>, a <dbl>,
#> #   b <dbl>, tz <dbl>, slope_tz <dbl>, fit <dbl>, fit_slope <dbl>,
#> #   start_z <dttm>