ploting fluxes for fit evaluation

plots the fluxes and indicates what should be discarded or replaced by zero

Usage

flux_plot_exp(
  slopes_df,
  datetime_col = "f_datetime",
  conc_col = "f_conc",
  cut_col = "f_cut",
  fit_col = "f_fit",
  fit_slope_col = "f_fit_slope",
  quality_flag_col = "f_quality_flag",
  fluxID_col = "f_fluxID",
  start_col = "f_start",
  b_col = "f_b",
  cor_coef_col = "f_cor_coef",
  RMSE_col = "f_RMSE",
  y_text_position = 500
)

Arguments

slopes_df

dataset containing slopes

datetime_col

column containing datetime of each concentration measurement

conc_col

column containing gas concentration data

cut_col

column containing cut factor from the flux_fitting function ("cut" or "keep")

fit_col

column containing the modelled fit of the flux

fit_slope_col

column containing the modelled slope at tz

quality_flag_col

column containing the flags produced by flux_quality

fluxID_col

column containing unique IDs for each flux

start_col

column containing the datetime of the start of each flux

b_col

column containing the b parameter of the exponential fit

cor_coef_col

column containing the correlation coefficient produced by flux_quality

RMSE_col

column containing the RMSE produced by flux_quality

y_text_position

position of the text box but will take time depending on the size of the dataset