ploting fluxes for fit evaluation

plots the fluxes and indicates what should be discarded or replaced by zero

Usage

flux_plot_lin(
  slopes_df,
  datetime_col = "f_datetime",
  conc_col = "f_conc",
  cut_col = "f_cut",
  fit_col = "f_fit",
  quality_flag_col = "f_quality_flag",
  fluxID_col = "f_fluxID",
  pvalue_col = "f_pvalue",
  rsquared_col = "f_rsquared",
  start_col = "f_start",
  y_text_position = 500
)

Arguments

slopes_df

dataset containing slopes

datetime_col

column containing datetime of each concentration measurement

conc_col

column containing gas concentration data

cut_col

column containing cut factor from the flux_fitting function ("cut" or "keep")

fit_col

column containing the modelled fit of the flux

quality_flag_col

column containing the flags produced by flux_quality

fluxID_col

column containing unique IDs for each flux

pvalue_col

column containing the p-value of each flux

rsquared_col

column containing the r squared used for the quality assessment

start_col

column containing the datetime of the start of each flux

y_text_position

position of the text box but will take time depending on the size of the dataset