quality assessment for the fluxes calculated with the linear model

indicates if fluxes should be discarded or replaced by 0 according to parameters set by user

Usage

flux_quality_lin(
  slopes_df,
  ambient_conc = 421,
  error = 100,
  fluxID_col = "f_fluxID",
  slope_col = "f_slope",
  conc_col = "f_conc",
  weird_fluxesID = c(),
  pvalue_col = "p.value",
  rsquared_col = "r.squared",
  pvalue_threshold = 0.3,
  rsquared_threshold = 0.7
)

Arguments

slopes_df

dataset containing slopes, fluxID, p.value and r.squared

ambient_conc

ambient gas concentration in ppm at the site of measurement (used to detect measurement that started with a polluted setup)

error

error of the setup, defines a window in which the starting values are considered acceptable

fluxID_col

column containing unique IDs for each flux

slope_col

column containing the slope of each flux (as calculated by the flux_fitting function)

conc_col

column containing the measured gas concentration

weird_fluxesID

vector of fluxIDs that should be discarded by the user's decision

pvalue_col

column containing the p-value of each flux

rsquared_col

column containing the r squared to be used for the quality assessment

pvalue_threshold

threshold of p-value below which the change of gas concentration over time is considered not significant (user decided)

rsquared_threshold

threshold of r squared value below which the linear model is considered an unsatisfactory fit

Value

same dataframe with added flag and corrected slopes columns

Examples

data(slopes0lin)
flux_quality_lin(slopes0lin,
  fluxID_col = "fluxID",
  slope_col = "slope", conc_col = "conc"
)
#> Joining with `by = join_by(f_fluxID)`
#> # A tibble: 1,251 × 24
#>    datetime            temp_air temp_soil f_conc   PAR turfID       type 
#>    <dttm>                 <dbl>     <dbl>  <dbl> <dbl> <fct>        <fct>
#>  1 2022-07-28 23:43:35    NA         NA     447. NA    156 AN2C 156 ER   
#>  2 2022-07-28 23:43:36     7.22      10.9   447.  1.68 156 AN2C 156 ER   
#>  3 2022-07-28 23:43:37    NA         NA     448. NA    156 AN2C 156 ER   
#>  4 2022-07-28 23:43:38    NA         NA     449. NA    156 AN2C 156 ER   
#>  5 2022-07-28 23:43:39    NA         NA     449. NA    156 AN2C 156 ER   
#>  6 2022-07-28 23:43:40    NA         NA     450. NA    156 AN2C 156 ER   
#>  7 2022-07-28 23:43:41    NA         NA     451. NA    156 AN2C 156 ER   
#>  8 2022-07-28 23:43:42    NA         NA     451. NA    156 AN2C 156 ER   
#>  9 2022-07-28 23:43:43    NA         NA     453. NA    156 AN2C 156 ER   
#> 10 2022-07-28 23:43:44    NA         NA     453. NA    156 AN2C 156 ER   
#> # ℹ 1,241 more rows
#> # ℹ 17 more variables: start <dttm>, end <dttm>, f_fluxID <fct>, n_conc <dbl>,
#> #   ratio <dbl>, flag <chr>, time <dbl>, cut <fct>, f_pvalue <dbl>,
#> #   f_rsquared <dbl>, adj.r.squared <dbl>, intercept <dbl>, f_slope <dbl>,
#> #   fit <dbl>, f_start_error <chr>, f_quality_flag <chr>, f_slope_corr <dbl>