quality assessment for the slopes estimated by flux_fitting

indicates if fluxes should be discarded or replaced by 0 according to parameters set by user. flux_quality_lm is for the model of the lm family. flux_quality_exp is for the exponential model.

Usage

flux_quality_lm(
  slopes_df,
  weird_fluxes_id = c(),
  force_ok_id = c(),
  pvalue_col = "f_pvalue",
  rsquared_col = "f_rsquared",
  pvalue_threshold = 0.3,
  rsquared_threshold = 0.7
)

Arguments

slopes_df

dataset containing slopes, fluxID, p.value and r.squared

weird_fluxes_id

vector of fluxIDs that should be discarded by the user's decision

force_ok_id

vector of fluxIDs for which the user wants to keep the calculated slope despite a bad quality flag

pvalue_col

column containing the p-value of each flux

rsquared_col

column containing the r squared to be used for the quality assessment

pvalue_threshold

threshold of p-value below which the change of gas concentration over time is considered not significant (user decided)

rsquared_threshold

threshold of r squared value below which the linear model is considered an unsatisfactory fit

Value

same dataframe with added flag and corrected slopes columns