Slopes for each flux

Slopes of C(t) for each flux with 30 seconds end cut, with quality flags.

Usage

slopes30_flag

Format

A tibble with 1251 rows and 36 variables

datetime

Datetime at which CO2 concentration was recorded.

temp_air

Air temperature inside the flux chamber in Celsius.

temp_soil

Ground temperature inside the flux chamber in Celsius.

f_conc

CO2 concentration in ppm.

PAR

Photosyntethically active radiation inside the chamber in micromol/s/sqm.

turfID

Unique ID of the turf in which the measurement took place.

type

Type of measurement: ecosystems respiration (ER) or net ecosystem exchange (NEE).

start

Datetime at which the measurement was started.

end

Datetime at which the measurement ended.

f_fluxID

Unique ID for each flux.

n_conc

Number of data point per flux.

ratio

Ratio of n_conc over length of the measurement (in seconds).

flag

Data quality flags.

f_time

Time variable of the flux in seconds.

f_cut

Indicating if the measurement should be kept (keep) or discarded (cut).

Cm_est

Estimation of the Cm parameter.

a_est

Estimation of the a parameter.

b_est

Estimation of the b parameter.

tz_est

Estimation of the tz parameter.

Cz

Cz parameter of the C(t) function.

Cm

Cm parameter of the C(t) function, calculated by optim() with Cm_est as starting point.

a

a parameter of the C(t) function, calculated by optim() with a_est as starting point.

f_b

b parameter of the C(t) function, calculated by optim() with b_est as starting point.

tz

tz parameter of the C(t) function, calculated by optim() with tz_est as starting point.

f_slope_tz

Slope of C(t) at tz

f_fit

C(t), modelled CO2 concentration as a function of time.

fit_slope

Output of linear model of CO2 concentration passing by C(tz) and a slope of slope_tz.

start_z

Datetime format of tz

f_cor_coef

coefficient of correlation between gas concentration and time

f_RMSE

RMSE of the exponential fit and the measured data

f_start_error

flagging if measurement started outside of the possible ambiant concentration

f_fit_quality

flagging bad fit

f_correlation

flagging if there is a correlation between gas concentrationa and time

f_quality_flag

quality flag advising if the slope has to be replaced by 0 or NA

f_slope_corr

slope corrected according to quality flag

Examples

slopes30_flag
#> # A tibble: 1,251 × 36
#>    f_datetime          temp_air temp_soil f_conc   PAR turfID       type 
#>    <dttm>                 <dbl>     <dbl>  <dbl> <dbl> <chr>        <chr>
#>  1 2022-07-28 23:43:35    NA         NA     447. NA    156 AN2C 156 ER   
#>  2 2022-07-28 23:43:36     7.22      10.9   447.  1.68 156 AN2C 156 ER   
#>  3 2022-07-28 23:43:37    NA         NA     448. NA    156 AN2C 156 ER   
#>  4 2022-07-28 23:43:38    NA         NA     449. NA    156 AN2C 156 ER   
#>  5 2022-07-28 23:43:39    NA         NA     449. NA    156 AN2C 156 ER   
#>  6 2022-07-28 23:43:40    NA         NA     450. NA    156 AN2C 156 ER   
#>  7 2022-07-28 23:43:41    NA         NA     451. NA    156 AN2C 156 ER   
#>  8 2022-07-28 23:43:42    NA         NA     451. NA    156 AN2C 156 ER   
#>  9 2022-07-28 23:43:43    NA         NA     453. NA    156 AN2C 156 ER   
#> 10 2022-07-28 23:43:44    NA         NA     453. NA    156 AN2C 156 ER   
#> # ℹ 1,241 more rows
#> # ℹ 29 more variables: f_start <dttm>, f_end <dttm>, f_fluxID <dbl>,
#> #   n_conc <dbl>, ratio <dbl>, flag <lgl>, f_time <dbl>, f_cut <chr>,
#> #   Cm_est <dbl>, a_est <dbl>, b_est <dbl>, tz_est <dbl>, f_Cz <dbl>,
#> #   time_diff <dbl>, f_Cm <dbl>, f_a <dbl>, f_b <dbl>, f_tz <dbl>,
#> #   f_slope_tz <dbl>, f_fit <dbl>, f_fit_slope <dbl>, f_start_z <dttm>,
#> #   f_cor_coef <dbl>, f_RMSE <dbl>, f_start_error <chr>, f_fit_quality <chr>, …