Slopes for each flux

Slopes of quadratic fit for each flux with 30 seconds end cut and t_zero of 10 seconds, with quality flags. C(t) = a + bt + ct^2

Usage

slopes30qua_flag

Format

A tibble with 1251 rows and 27 variables

f_datetime

Datetime at which CO2 concentration was recorded.

temp_air

Air temperature inside the flux chamber in Celsius.

temp_soil

Ground temperature inside the flux chamber in Celsius.

f_conc

CO2 concentration in ppm.

PAR

Photosynthetically active radiation inside the chamber in micromol/s/sqm.

turfID

Unique ID of the turf in which the measurement took place.

type

Type of measurement: ecosystems respiration (ER) or net ecosystem exchange (NEE).

f_start

Datetime at which the measurement was started.

f_end

Datetime at which the measurement ended.

f_fluxID

Unique ID for each flux.

n_conc

Number of data point per flux.

ratio

Ratio of n_conc over length of the measurement (in seconds).

flag

Data quality flags.

f_time

Time variable of the flux in seconds.

f_cut

Indicating if the measurement should be kept (keep) or discarded (cut).

f_pvalue

P-value of the quadratic model of gas concentration over time.

f_rsquared

R squared of the quadratic model of gas concentration over time.

f_adj_rsquared

Adjusted R squared of the quadratic model of gas concentration over time.

f_intercept

Intercept of the quadratic model of gas concentration over time.

f_param1

b parameter of C(t)

f_param2

c parameter of C(t)

f_slope

Slope of the quadratic model of gas concentration over time at t_zero.

f_fit

Output of the quadratic model of gas concentration over time.

f_fit_slope

output of linear expression describing the slope at t_zero

f_start_error

flagging if measurement started outside of the possible ambient concentration

f_quality_flag

quality flag advising if the slope has to be replaced by 0 or NA

f_slope_corr

slope corrected according to quality flag

Examples

slopes30qua_flag
#> # A tibble: 1,251 × 30
#>    f_datetime          temp_air temp_soil f_conc   PAR turfID       type 
#>    <dttm>                 <dbl>     <dbl>  <dbl> <dbl> <fct>        <fct>
#>  1 2022-07-28 23:43:35    NA         NA     447. NA    156 AN2C 156 ER   
#>  2 2022-07-28 23:43:36     7.22      10.9   447.  1.68 156 AN2C 156 ER   
#>  3 2022-07-28 23:43:37    NA         NA     448. NA    156 AN2C 156 ER   
#>  4 2022-07-28 23:43:38    NA         NA     449. NA    156 AN2C 156 ER   
#>  5 2022-07-28 23:43:39    NA         NA     449. NA    156 AN2C 156 ER   
#>  6 2022-07-28 23:43:40    NA         NA     450. NA    156 AN2C 156 ER   
#>  7 2022-07-28 23:43:41    NA         NA     451. NA    156 AN2C 156 ER   
#>  8 2022-07-28 23:43:42    NA         NA     451. NA    156 AN2C 156 ER   
#>  9 2022-07-28 23:43:43    NA         NA     453. NA    156 AN2C 156 ER   
#> 10 2022-07-28 23:43:44    NA         NA     453. NA    156 AN2C 156 ER   
#> # ℹ 1,241 more rows
#> # ℹ 23 more variables: f_start <dttm>, f_end <dttm>, f_fluxID <fct>,
#> #   n_conc <int>, ratio <dbl>, flag <chr>, f_time <dbl>, f_cut <fct>,
#> #   f_rsquared <dbl>, f_adj_rsquared <dbl>, f_pvalue <dbl>, f_intercept <dbl>,
#> #   f_param1 <dbl>, f_param2 <dbl>, f_slope <dbl>, f_fit <dbl>,
#> #   f_fit_slope <dbl>, f_n_conc <int>, f_ratio <dbl>, f_flag_ratio <chr>,
#> #   f_start_error <chr>, f_quality_flag <chr>, f_slope_corr <dbl>