calculates fluxes

calculates a flux based on the rate of change of gas concentration over time

Usage

flux_calc(
  slope_df,
  slope_col,
  chamber_volume = 24.5,
  tube_volume = 0.075,
  atm_pressure = 1,
  plot_area = 0.0625,
  R_const = 0.082057,
  cols_keep = c(),
  cols_ave = c(),
  fluxID_col = "fluxID",
  temp_air_col = "temp_air",
  temp_air_unit = "celsius"
)

Arguments

slope_df

dataframe of flux slopes

slope_col

column containing the slope to calculate the flux

chamber_volume

volume of the flux chamber in L, default for Three-D project chamber (25x24.5x40cm)

tube_volume

volume of the tubing in L, default for summer 2020 setup

atm_pressure

atmoshperic pressure, assumed 1 atm

plot_area

area of the plot in m^2, default for Three-D

R_const

gas constant, in LatmK^(-1)*mol^(-1)

cols_keep

columns to keep from the input to the output. Those columns need to have unique values for each flux.

cols_ave

columns with values that should be averaged for each flux in the ouput. Note that NA are removed in mean calculation

fluxID_col

column containing the fluxID

temp_air_col

column containing the air temperature used to caculate fluxes. Will be averaged with NA removed.

temp_air_unit

units in which air temperature was measured. Has to be either celsius, fahrenheit or kelvin

Value

a df containing fluxID, fluxes, temperature average for each flux, slope used for each flux calculation, and any columns specified in cols_keep and cols_ave

Examples

data(slopes0)
flux_calc(slopes0, slope_col = "slope_tz")
#> # A tibble: 6 × 4
#>   fluxID slope temp_air_ave  flux
#>    <dbl> <dbl>        <dbl> <dbl>
#> 1      1 1.56          7.31  95.6
#> 2      2 0.853         7.38  52.4
#> 3      3 0.303         7.46  18.6
#> 4      4 1.13          7.77  69.4
#> 5      5 1.46          7.71  89.9
#> 6      6 0.426         7.75  26.2