quadratic fit to gas concentration over time

fits a quadratic model to the gas concentration over time

Usage

flux_fitting_quadratic(
  conc_df,
  start_cut = 0,
  end_cut = 0,
  start_col = "f_start",
  end_col = "f_end",
  datetime_col = "f_datetime",
  conc_col = "f_conc",
  fluxid_col = "f_fluxID",
  t_zero = 0
)

Arguments

conc_df

dataframe of gas concentration over time

start_cut

time to discard at the start of the measurements (in seconds)

end_cut

time to discard at the end of the measurements (in seconds)

start_col

column with datetime when the measurement started

end_col

column with datetime when the measurement ended

datetime_col

column with datetime of each concentration measurement

conc_col

column with gas concentration data

fluxid_col

column with ID of each flux

t_zero

time at which the slope should be calculated

Value

a df with the modeled gas concentration, slope, intercept, std error, r square and p value of the quadratic model