quality assessment for the slopes estimated by flux_fitting

indicates if fluxes should be discarded or replaced by 0 according to parameters set by user. flux_quality_lm is for the model of the lm family. flux_quality_exp is for the exponential model.

Usage

flux_quality_exp(
  slopes_df,
  force_discard = c(),
  force_ok = c(),
  b_col = "f_b",
  rmse_threshold = 25,
  cor_threshold = 0.5,
  b_threshold = 1
)

Arguments

slopes_df

dataset containing slopes, fluxID, and parameters of the exponential expression

force_discard

vector of fluxIDs that should be discarded by the user's decision

force_ok

vector of fluxIDs for which the user wants to keep the calculated slope despite a bad quality flag

b_col

column containing the b parameter of the exponential expression

rmse_threshold

threshold for the RMSE of each flux above which the fit is considered unsatisfactory

cor_threshold

threshold for the correlation coefficient of gas concentration with time below which the correlation is considered non significant

b_threshold

threshold for the b parameter. Defines a window with its opposite inside which the fit is considered good enough.

Value

same dataframe with added flag and corrected slopes columns