assessing fluxes quality calculated with an exponential fit quality assessment for the fluxes calculated with the exponential model

indicates if fluxes should be discarded or replaced by 0 according to parameters set by user

Usage

flux_quality_exp(
  slopes_df,
  ambient_conc = 421,
  error = 100,
  fluxID_col = "f_fluxID",
  slope_col = "f_slope",
  weird_fluxesID = c(),
  conc_col = "f_conc",
  b_col = "f_b",
  time_col = "f_time",
  fit_col = "f_fit",
  cut_col = "f_cut",
  RMSE_threshold = 25,
  cor_threshold = 0.5,
  b_threshold = 1
)

Arguments

slopes_df

dataset containing slopes, fluxID, and parameters of the exponential expression

ambient_conc

ambient gas concentration in ppm at the site of measurement (used to detect measurement that started with a polluted setup)

error

error of the setup, defines a window in which the starting values are considered acceptable

fluxID_col

column containing unique IDs for each flux

slope_col

column containing the slope of the exponential expression used for the calculation

weird_fluxesID

vector of fluxIDs that should be discarded by the user's decision

conc_col

column containing the measured gas concentration

b_col

column containing the b parameter of the exponential expression

time_col

column containing the time of each measurement in seconds

fit_col

column containing the modeled data

cut_col

column containing the cutting information

RMSE_threshold

threshold for the RMSE of each flux above which the fit is considered unsatisfactory

cor_threshold

threshold for the correlation coefficient of gas concentration with time below which the correlation is considered non significant

b_threshold

threshold for the b parameter. Defines a window with its opposite inside which the fit is considered good enough.

Value

same dataframe with added flag and corrected slopes columns